requirements.txt
yfinance>=0.2.0
google-generativeai>=0.3.0
pytz>=2024.1
